Move LoginUtil to common security package
